\name{evt0-package}
\alias{evt0-package}
\alias{evt0}
\docType{package}
\title{
Mean of Order P, Peaks over Random Threshold Hill and High Quantile Estimates
}
\description{
Computes extreme value index (EVI) estimate for heavy tailed models by mean of order p (MOP) 
and peaks over random threshold (PORT) Hill methodologies. Besides, also computes moment, generalised Hill and mixed moment estimates for EVI. Compute high quantile or value-at-risk (VaR) based on above EVI estimates.
}

\author{
B G Manjunath \email{bgmanjunath@gmail.com} and Frederico Caeiro \email{fac@fct.unl.pt}; guidance from Prof. M. Ivette Gomes \email{ivette.gomes@fc.ul.pt} and Prof. M. Isabel Fraga Alves \email{mialves@fc.ul.pt}
}
