\name{evt0-package}
\alias{evt0-package}
\alias{evt0}
\docType{package}
\title{
Mean of order p (MOP) and peaks over random threshold (PORT) Hill estimates for the extreme value index (EVI)
}
\description{
Compute EVI estimate using MOP and PORT methodology. 
}
\details{
\tabular{ll}{
Package: \tab evt0\cr
Type: \tab Package\cr
Version: \tab 1.1-0\cr
Date: \tab 2013-11-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
B G Manjunath \cr
Maintainer: B G Manjunath <bgmanjunath@gmail.com>
}
