\name{evt0-package}
\alias{evt0-package}
\alias{evt0}
\docType{package}
\title{
Mean of order p, peaks over random threshold Hill and high quantile estimates
}
\description{
Computes extreme value index (EVI) estimate for heavy tailed models by Mean of order p (MOP) 
and peaks over random threshold (PORT) Hill methodologies. Besides, also computes moment, generalised Hill and mixed moment estimates for EVI. Compute high quantile or value-at-risk (VaR) based on above EVI estimates.
}
\details{
\tabular{ll}{
Package: \tab evt0\cr
Type: \tab Package\cr
Version: \tab 1.1-3\cr
Date: \tab 2013-12-10\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
B G Manjunath \email{bgmanjunath@gmail.com} and Frederico Caeiro \email{fac@fct.unl.pt}; guidance from Prof. M. Ivette Gomes \email{ivette.gomes@fc.ul.pt} and Prof. M. Isabel Fraga Alves \email{mialves@fc.ul.pt}


Maintainer: B G Manjunath \email{bgmanjunath@gmail.com}
}
